requirements.txt
asgiref==3.11.0
black==25.11.0
click==8.3.1
colorama==0.4.6
Django==5.2.8
djangorestframework==3.16.1
djangorestframework-simplejwt>=5.3,<6.0
django-filter>=24.2,<25.0
flake8==7.3.0
isort==7.0.0
mccabe==0.7.0
mypy==1.18.2
mypy_extensions==1.1.0
packaging==25.0
pathspec==0.12.1
pillow==12.0.0
platformdirs==4.5.0
psycopg2-binary==2.9.11
pycodestyle==2.14.0
pyflakes==3.4.0
python-dotenv==1.2.1
pytokens==0.3.0
setuptools==80.9.0
sqlparse==0.5.3
typing_extensions==4.15.0
tzdata==2025.2
wheel==0.45.1
